Refactor Video component
